Refactor: Move passkey authentication routes
