+++ app/AppComponents.scala
@@ -128,6 +128,16 @@ class AppComponents(context: ApplicationLoader.Context)
       passkeysEnablingCookieName,
       passkeyAuthenticatorMetadata
     ),
+    new PasskeyController(
+      controllerComponents,
+      authAction,
+      passkeyAuthAction,
+      passkeyRegistrationAuthAction,
+      host,
+      janusData,
+      passkeysEnabled,
+      passkeysEnablingCookieName
+    ),
     new Audit(janusData, controllerComponents, authAction),
     new RevokePermissions(
       janusData,
@@ -143,20 +153,11 @@ class AppComponents(context: ApplicationLoader.Context)
       googleGroupChecker,
       requiredGoogleGroups
     ),
-    new PasskeyController(
-      controllerComponents,
-      authAction,
-      passkeyAuthAction,
-      passkeyRegistrationAuthAction,
-      host,
-      janusData
-    ),
     new Utility(
       janusData,
       controllerComponents,
       authAction,
-      configuration,
-      passkeysEnablingCookieName
+      configuration
     ),
     assets
   )
