+++ conf/routes
@@ -7,11 +7,17 @@ GET     /                           controllers.Janus.index
 GET     /superuser                  controllers.Janus.admin
 GET     /support                    controllers.Janus.support
 GET     /user-account               controllers.Janus.userAccount
-POST    /console                    controllers.Janus.consoleLogin(permissionId: String)
-POST    /consoleUrl                 controllers.Janus.consoleUrl(permissionId: String)
-POST    /credentials                controllers.Janus.credentials(permissionId: String)
-POST    /multi-credentials          controllers.Janus.multiCredentials(permissionIds: String)
 POST    /favourite                  controllers.Janus.favourite()
+
+# Passkey-protected routes
+GET     /console                    controllers.PasskeyController.showAuthPage
+POST    /console                    controllers.Janus.consoleLogin(permissionId: String)
+GET     /consoleUrl                 controllers.PasskeyController.showAuthPage
+POST    /consoleUrl                 controllers.Janus.consoleUrl(permissionId: String)
+GET     /credentials                controllers.PasskeyController.showAuthPage
+POST    /credentials                controllers.Janus.credentials(permissionId: String)
+GET     /multi-credentials          controllers.PasskeyController.showAuthPage
+POST    /multi-credentials          controllers.Janus.multiCredentials(permissionIds: String)
 
 GET     /audit/user/:username       controllers.Audit.byUser(username: String)
 POST    /audit/user/:username       controllers.Audit.changeUserDate(username: String)
@@ -35,17 +41,11 @@ POST    /passkey/register                   controllers.PasskeyController.regist
 DELETE  /passkey/:passkeyId                 controllers.PasskeyController.deletePasskey(passkeyId: String)
 
 # Tmp routes for passkey trial
-GET     /opt/in/passkeys            controllers.Utility.optInToPasskeys
-GET     /opt/out/passkeys           controllers.Utility.optOutOfPasskeys
+GET     /opt/in/passkeys            controllers.PasskeyController.optInToPasskeys
+GET     /opt/out/passkeys           controllers.PasskeyController.optOutOfPasskeys
 
 GET     /healthcheck                controllers.Utility.healthcheck
 GET     /accounts                   controllers.Utility.accounts
-
-# Old routes that moved to POST when passkey authentication was introduced
-GET     /console                    controllers.Utility.gone
-GET     /consoleUrl                 controllers.Utility.gone
-GET     /credentials                controllers.Utility.gone
-GET     /multi-credentials          controllers.Utility.gone
 
 # Map static resources from the /public folder to the /assets URL path
 GET     /assets/*file               controllers.Assets.versioned(file)
